stage1_ocr/requirements.txt
numpy==2.4.4
tqdm==4.67.3
torch==2.11.0
pillow==12.2.0
matplotlib==3.10.8
torchvision==0.26.0
opencv-python>=4.10
optuna>=3.6
scikit-learn>=1.4
fastapi==0.136.0
pydantic==2.13.2
uvicorn==0.44.0
